머지후코드정리 + 포션 기능 수정
